Renamed `/rcc-suggest` to `/rcc-improve`
