+++ ai-assistant/src/RocketChatterApp.ts
@@ -22,8 +22,8 @@ import { DocumentCommand } from "./commands/DocumentCommand";
 import { FindSimilarCommand } from "./commands/FindSimilar";
 import { HelpCommand } from "./commands/HelpCommand";
 import { ImportanceCommand } from "./commands/ImportanceCommand";
+import { ImproveCommand } from "./commands/ImproveCommand";
 import { StyleguideCommand } from "./commands/Styleguide";
-import { SuggestCommand } from "./commands/SuggestCommand";
 import { TranslateCommand } from "./commands/TranslateCommand";
 import { WhyUsedCommand } from "./commands/WhyUsedCommand";
 import { EstablishRelationsEndpoint } from "./endpoints/establishRelations";
@@ -63,7 +63,7 @@ export class RocketChatterApp extends App {
         configuration.slashCommands.provideSlashCommand(
             new StyleguideCommand()
         );
-        configuration.slashCommands.provideSlashCommand(new SuggestCommand());
+        configuration.slashCommands.provideSlashCommand(new ImproveCommand());
         configuration.slashCommands.provideSlashCommand(new TranslateCommand());
         configuration.slashCommands.provideSlashCommand(new WhyUsedCommand());
