+++ ai-assistant/src/core/prompt/prompt.factory.ts
@@ -2,7 +2,6 @@ import { PROMPT_DIAGRAM_COMMAND } from "./contents/diagram";
 import { PROMPT_DOCUMENT_COMMAND } from "./contents/document";
 import { PROMPT_EXTRACT_DB_KEYWORDS } from "./contents/extractDBKeywords";
 import { PROMPT_STYLEGUIDE_COMMAND } from "./contents/styleguide";
-import { PROMPT_SUGGEST_COMMAND } from "./contents/suggest";
 import { PROMPT_TRANSLATE_COMMAND } from "./contents/translate";
 import { PROMPT_WHY_USED_COMMAND } from "./contents/whyUsed";
 import { Prompt } from "./prompt";
@@ -93,13 +92,39 @@ export namespace PromptFactory {
         return prompt;
     }
 
-    export function makeSuggestPrompt(
+    export function makeImprovePrompt(
         codebase: string,
         targetEntity: string
     ): Prompt {
         const prompt = new Prompt();
 
-        prompt.pushSystem(PROMPT_SUGGEST_COMMAND);
+        prompt.pushSystem(`
+            You are an expert in understanding typescript and javascript codebases and fixing it provided the context of the codebase.
+
+            INPUT: Other entities the target entity might be using. The target entity to refactor.
+
+            TASKS:
+            - Refactoring might include:
+            - Renaming
+            - Extracting different parts into separate functions
+            - Making code concise to make it more readable, maintainable
+            - Removing dead code
+            - Performance improvements
+            - Better alternatives
+            - Syntax improvements
+            - Code style improvements
+            - Best practices
+            - Suggest multiple (only if relevant) fixes for the target entity.
+            - If the target entity is already correct then tell that it is already correct.
+            - If the provided codebase contains entities that are functionally similar to what's used in the target entity, suggest using entities from the codebase.
+
+            EXPECTED OUTPUT: Suggestions for the target entity in form of MARKDOWN and CODE SNIPPET with the fix and explanation.
+
+            RULES:
+            - STRICTLY, do not make anything other than the answer to the user's query.
+            - Do not provide any kind of diagram or visualization in the output.
+            - The output MUST BE IN ONLY AND ONLY MARKDOWN.
+        `);
         prompt.pushUser(
             `Hey, can you suggest multiple fixes for the target entity? To help you with the context I have provided the codebase of the entities it uses and the target entity. You don't need to worry about the codebase, just focus on the target entity.\n\n<CODEBASE_START>\n${codebase}\n<CODEBASE_END>\n<TARGET_ENTITY_START>${targetEntity}\n<TARGET_ENTITY_END>.`
         );
